requirements.txt
argparse
requests
pyyaml
dataclasses
uuid
pytest